+++ README.md
@@ -172,6 +172,7 @@
 - [x] 헤더 모달 활성화 버튼
 - [ ] 모달 컴포넌트 제작
   - [x] 모달 title 컴포넌트 분리
+    - [x] 모달 title 컴포넌트 리팩토링
   - [ ] 모달 프레임 컴포넌트 제작
 - [ ] E2E 테스트 진행
 - [ ] 리팩토링
